+++ mobilemenu/appinfo/routes.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+	'routes' => [
+		[
+			'name' => 'settings#save',
+			'url' => '/settings',
+			'verb' => 'POST',
+		],
+	],
+];